requirements.txt
matplotlib
pandas
regex
python-dotenv
openai
